Install plugins only after switching vim symlinks
Ideally we want to swicth to the new config only after all the plugins are
installed successfully. But there seems to be no way to do that since
Plug wants to install them in ~/.vim, I do not know of a way to ask
it to install them in HVN_DEST/.vim instead.