Refactor to allow a basic setup without Haskell components (#183)
* Support installation without haskell components

Refactor to separate the haskell setup from core vim and plugins setup so that
we can optionally install the rest of the config without haskell components.

* Make error handling more robust

* Update README with advanced install options

* Update stack install instructions' URL

* Install plugins only after switching vim symlinks

Ideally we want to swicth to the new config only after all the plugins are
installed successfully. But there seems to be no way to do that since
Plug wants to install them in ~/.vim, I do not know of a way to ask
it to install them in HVN_DEST/.vim instead.